backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.4
gunicorn==20.1.0
google-genai==1.3.0
sentence-transformers
faiss-cpu
numpy